added alert

+++ components/Favorites.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, FlatList, Text } from 'react-native';
+import { View, FlatList, Text, Alert} from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseURL } from '../shared/baseURL';
@@ -32,7 +32,26 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    /* directly deleting: onPress: () => this.props.deleteFavorite(item.id) */
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite',
+                            'Are you sure wish to delete the favorite dish ' + item.name + '?',
+                            // Alert Buttons
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log("Not deleted.."),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false } // Must press OK or Cancel
+                        )
+                    }
                 }
             ]
             return(
